[major] move the classes UpConversionProtocol, DownConversionProtocol, EmittingProtocol from the example folder to the core code base; fix the unit tests
