+++ sequence/components/transmon.py
@@ -3,12 +3,16 @@
 See https://arxiv.org/abs/2411.11377, Simulation of Quantum Transduction Strategies for Quantum Networks
 """
 
+import random
 from typing import List
 import numpy as np
 from ..kernel.entity import Entity
 from ..kernel.timeline import Timeline
 from ..topology.node import Node
+from ..protocol import Protocol
 from .photon import Photon
+from .transducer import Transducer
+from sequence.constants import KET1
 
 
 class Transmon(Entity):
@@ -41,7 +45,6 @@ class Transmon(Entity):
         self.efficiency = efficiency
         self.input_quantum_state = None
               
-
     def init(self):
         pass
 
@@ -53,13 +56,11 @@ class Transmon(Entity):
 
     def generation(self):
         """Set tranmson quantum state and return a photon with some specifications"""
-        
         input_quantum_state = np.kron(self.photons_quantum_state[0], self.photons_quantum_state[1])
         self.input_quantum_state = input_quantum_state
 
         return Photon(name="photon", timeline=self.timeline, wavelength=self.wavelengths[0], quantum_state=self.photons_quantum_state[0])
         
-
     def get(self, photon: Photon) -> None:
         """Receive photon from the transducer
         
@@ -70,6 +71,43 @@ class Transmon(Entity):
         print(f"Photon received by the Transmon at Rx: {photon}, Name: {photon.name}, Wavelength: {photon.wavelength}")
         print(f"Photon counter of TRANSMON {self.photon_counter}")
 
-
     def receive_photon(self, photon: Photon) -> None:
         self.photon_counter += 1
+
+
+class EmittingProtocol(Protocol):
+    """Protocol for emission of single microwave photon by transmon.
+
+    Attributes:
+        owner (Node): the owner of this protocol, the protocol runs on the owner
+        name (str): the name of the protocol
+        tl (Timeline): the simulation timeline
+        transmon (Transmon): the transmon component
+        transducer (Transducer): the transducer component
+    """
+
+    def __init__(self, owner: "Node", name: str, tl: Timeline, transmon: Transmon, transducer: Transducer):
+        super().__init__(owner, name)
+        self.owner = owner
+        self.name = name
+        self.tl = tl
+        self.transmon = transmon
+        self.transducer = transducer
+
+    def start(self):
+        print(f"EmittingProtocol started for {self.owner.name} at time {self.tl.now()}")
+        photon = self.transmon.generation()
+        print(f"Photon created: {photon}, Name: {photon.name}, Wavelength: {photon.wavelength}")
+        print(f"Transmon at Tx quantum state: {self.transmon.input_quantum_state} of {self.owner.name}")
+
+        if self.transmon.photons_quantum_state[0] == KET1:
+            if random.random() < self.transmon.efficiency:
+                print(f"Transmon receiver " + str(self.transmon._receivers[0]))
+                self.transmon._receivers[0].receive_photon_from_transmon(photon)
+            else:
+                print("Photon emission failed due to transmon efficiency")
+        else:
+            print("The transmon is in the state 00, or 01, it doesn't emit microwave photons")
+        
+    def received_message(self, src: str, msg):
+        pass
